requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
pydantic==2.6.4
python-dotenv==1.0.1